Revert "новые файлы Модифицированного калькулятора"
This reverts commit 39c666dff9eda1ea3037c86648121bb8c54fe040.

+++ Lesson_1/Calculator.java
@@ -1,9 +1,9 @@
 public class Calculator {
     public static void main(String[] args) {
-        double num1 = 1423L;
-        double num2 = 17;
+        long num1 = 1423L;
+        long num2 = 17;
         char sign = '^'; // здесь можно задать нужный знак операции
-        double result;
+        long result;
 
         if (sign == '+') {
             result = num1 + num2;
